Include packages in repository
I'm going to set up CircleCI for CI builds for this repository once I
publish it on GitHub, and I've found that installing NuGet to restore the
packages is hard to do from a Linux script -- in fact, I haven't found any
way to do so. I found that you can manually download the NuGet executable
and upload it to your server, but that isn't what I want.