Tweak FilesBetween to simplify CanCastle
Because I feel that the "not is King" constraint could cause trouble in
atomic when the opposite king is on the same rank as the castling king.
Although I haven't been able to find a case where this was actually a
problem, I didn't trust that enough to be sure - and that aside, I feel
like the way FilesBetween works now makes more sense anyway.